Store session state in Cassandra using containers
As a user I want to be able to resume a survey on a different device. To enable this
the system needs to store state before it is submitted downstream. We have choosen to
test Cassandra as a Key / Value store for this session data. This commit creates the
Cassandra schema (keyspace and table) via the manage.py script which is idempotent and
called everytime a container is started. The cassandra schema is created intentionally
simple, using simply a flask session ID (created using UUID), the questionnaire session ID
(created by the respondent system) and the session data.

We make use of a new auto save method that takes an AJAX post of the form values, serialised
to json in a format that WTForms can use for initial data. When the questionnaire session is
reloaded, the data is pulled from cassandra into the initial form data for rendering to the
user.

We have implemented docker-compose, as a local orchestration tool to allow our dev team to
work with cassandra and flask locally, replicating our production environment. In eq-terraform,
we have a similar piece of work that creates an AWS elastic container service cluster, using
the same Dockerfile as used by docker-compose.

We also have created a dummy survey schema that can be used for development without having
to standup a eq-author instance - this is designed to speed up dev and pay down technical debt
referenced in DIGEQ-106. As we enhance our survey schema design this will need updating.