DIGEQ-56 Respondent view rich text form field
This commit creates a RichTextField that enables an Author to create
a text box in their questionnaire to inform a respondent.

We've added a new convertor to jsonforms and created a whitelist
approach to HTML tags based on the python bleach package.

This commit also creates a folder for test fixtures so that we can
create a range of new survey examples without having to have the
strings inline in the test files.

The requirements have been updated and will require a rebuild of
our downstream docker images.

We created a custom widget for displaying a text block.